src/e_cidadania/apps/admin/views.py
# -*- coding: utf-8 -*-
#
# Copyright (c) 2010 Cidadanía Coop.
# Written by: Oscar Carballal Prego <info@oscarcp.com>
#
# This file is part of e-cidadania.
#
# e-cidadania is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# e-cidadania is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with e-cidadania. If not, see <http://www.gnu.org/licenses/>.


import datetime

from django.views.generic.base import TemplateView, RedirectView
from django.views.generic.list import ListView
from django.views.generic.edit import CreateView, UpdateView, DeleteView
from django.views.generic.detail import DetailView

from django.contrib.auth.models import User, Group
from django.contrib import messages
from django.template import RequestContext

from e_cidadania.apps.admin.models import Admin

class AdminIndex(DetailView):
    
    """
    Serves the administration index page, containing all the configuration
    settings for the modules and other stuff.
    """
    context_object_name = 'admin'
    template_name = 'admin/admin_index.html'
    
    def get_object(self):
        adminobj = get_object_or_404(Admin)