src/e_cidadania/requirements.txt
python-pil
django-tagging
django-wysiwyg
django-grappelli
feedparser
python-dateutil==1.5
